+++ components/collector/tests/source_collectors/api_source_collectors/test_jira.py
@@ -10,6 +10,8 @@ from tests.source_collectors.source_collector_test_case import SourceCollectorTe
 class JiraTestCase(SourceCollectorTestCase):
     """Base class for Jira unit tests."""
 
+    METRIC_TYPE = "Subclass responsibility"
+
     def setUp(self):
         super().setUp()
         self.sources = dict(
@@ -19,36 +21,45 @@ class JiraTestCase(SourceCollectorTestCase):
                     url="https://jira/", jql="query", story_points_field="field",
                     manual_test_execution_frequency_field="desired_test_frequency",
                     manual_test_duration_field="field")))
-        self.fields_json = []
+        self.metric = dict(type=self.METRIC_TYPE, addition="sum", sources=self.sources)
+        self.created = "2020-08-06T16:36:48.000+0200"
+
+    def issue(self, key="1", **fields):
+        """Create a Jira issue."""
+        return dict(id=key, key=key, fields=dict(created=self.created, summary=f"Summary {key}", **fields))
+
+    def entity(self, key="1", created=None, updated=None, **kwargs):
+        """Create an entity."""
+        return dict(
+            key=key, summary=f"Summary {key}", url=f"https://jira/browse/{key}", created=created or self.created,
+            updated=updated, status=None, priority=None, **kwargs)
+
+    async def get_response(self, issues_json, fields_json=None):
+        return await self.collect(self.metric, get_request_json_side_effect=[fields_json or [], issues_json])
 
 
 class JiraIssuesTest(JiraTestCase):
     """Unit tests for the Jira issue collector."""
 
-    def setUp(self):
-        super().setUp()
-        self.metric = dict(type="issues", addition="sum", sources=self.sources)
+    METRIC_TYPE = "issues"
 
     async def test_nr_of_issues(self):
         """Test that the number of issues is returned."""
         issues_json = dict(total=42)
-        response = await self.collect(self.metric, get_request_json_side_effect=[self.fields_json, issues_json])
+        response = await self.get_response(issues_json)
         self.assert_measurement(response, value="42")
 
     async def test_issues(self):
         """Test that the issues are returned."""
-        issues_json = dict(total=1, issues=[dict(key="key", id="id", fields=dict(summary="Summary"))])
-        response = await self.collect(self.metric, get_request_json_side_effect=[self.fields_json, issues_json])
-        self.assert_measurement(response, entities=[dict(key="id", summary="Summary", url="https://jira/browse/key")])
+        issues_json = dict(total=1, issues=[self.issue()])
+        response = await self.get_response(issues_json)
+        self.assert_measurement(response, value="1", entities=[self.entity()])
 
 
 class JiraManualTestExecutionTest(JiraTestCase):
     """Unit tests for the Jira manual test execution collector."""
 
-    def setUp(self):
-        super().setUp()
-        self.too_long_ago_summary = "Tested too long ago according to its own desired frequency"
-        self.metric = dict(type="manual_test_execution", addition="sum", sources=self.sources)
+    METRIC_TYPE = "manual_test_execution"
 
     async def test_nr_of_test_cases(self):
         """Test that the number of test cases is returned."""
@@ -56,79 +67,54 @@ class JiraManualTestExecutionTest(JiraTestCase):
         ten_days_ago = str(datetime.now()-timedelta(days=10))
         test_cases_json = dict(
             issues=[
-                dict(key="key1", id="id1",
-                     fields=dict(comment=dict(comments=[dict(updated=long_ago)]), summary="Tested too long ago")),
-                dict(key="key2", id="id2", fields=dict(comment=dict(comments=[]), summary="Never tested")),
-                dict(key="key3", id="id3",
-                     fields=dict(
-                         comment=dict(comments=[dict(updated=str(datetime.now()))]), summary="Recently tested")),
-                dict(key="key4", id="id4",
-                     fields=dict(
-                         comment=dict(comments=[dict(updated=ten_days_ago)]), desired_test_frequency="5",
-                         summary=self.too_long_ago_summary))])
-        response = await self.collect(self.metric, get_request_json_side_effect=[self.fields_json, test_cases_json])
+                self.issue(key="1", comment=dict(comments=[dict(updated=long_ago)])),
+                self.issue(key="2", comment=dict(comments=[])),
+                self.issue(key="3", comment=dict(comments=[dict(updated=str(datetime.now()))])),
+                self.issue(key="4", comment=dict(comments=[dict(updated=ten_days_ago)]), desired_test_frequency="5")])
+        response = await self.get_response(test_cases_json)
         self.assert_measurement(
             response, value="2",
             entities=[
-                dict(key="id1", summary="Tested too long ago", url="https://jira/browse/key1",
-                     last_test_date=str(parse(long_ago).date()), desired_test_frequency="21"),
-                dict(key="id4", summary=self.too_long_ago_summary, url="https://jira/browse/key4",
-                     last_test_date=str(parse(ten_days_ago).date()), desired_test_frequency="5")])
+                self.entity(key="1", last_test_date=str(parse(long_ago).date()), desired_test_frequency="21"),
+                self.entity(key="4", last_test_date=str(parse(ten_days_ago).date()), desired_test_frequency="5")])
 
     async def test_nr_of_test_cases_with_field_name(self):
         """Test that the number of test cases is returned when the field name for the test frequency is specified
         by name."""
-        fields_json = [dict(name="Required test frequency", id="custom_field_001")]
         self.sources["source_id"]["parameters"]["manual_test_execution_frequency_field"] = "Required test frequency"
         ten_days_ago = str(datetime.now()-timedelta(days=10))
         test_cases_json = dict(
-            issues=[
-                dict(key="key", id="id",
-                     fields=dict(
-                         comment=dict(comments=[dict(updated=ten_days_ago)]), custom_field_001="5",
-                         summary=self.too_long_ago_summary))])
-        response = await self.collect(self.metric, get_request_json_side_effect=[fields_json, test_cases_json])
+            issues=[self.issue(comment=dict(comments=[dict(updated=ten_days_ago)]), custom_field_001="5")])
+        fields_json = [dict(name="Required test frequency", id="custom_field_001")]
+        response = await self.get_response(test_cases_json, fields_json)
         self.assert_measurement(
             response, value="1",
-            entities=[
-                dict(key="id", summary=self.too_long_ago_summary, url="https://jira/browse/key",
-                     last_test_date=str(parse(ten_days_ago).date()), desired_test_frequency="5")])
+            entities=[self.entity(key="1", last_test_date=str(parse(ten_days_ago).date()), desired_test_frequency="5")])
 
 
 class JiraReadyUserStoryPointsTest(JiraTestCase):
     """Unit tests for the Jira ready story points collector."""
 
+    METRIC_TYPE = "ready_user_story_points"
+
     async def test_nr_story_points(self):
         """Test that the number of story points is returned."""
-        metric = dict(type="ready_user_story_points", addition="sum", sources=self.sources)
-        user_stories_json = dict(
-            issues=[
-                dict(key="1", id="1", fields=dict(summary="user story 1", field=10)),
-                dict(key="2", id="2", fields=dict(summary="user story 2", field=32))])
-        response = await self.collect(
-            metric, get_request_json_side_effect=[self.fields_json, user_stories_json])
+        user_stories_json = dict(issues=[self.issue(key="1", field=10), self.issue(key="2", field=32)])
+        response = await self.get_response(user_stories_json)
         self.assert_measurement(
-            response, value="42",
-            entities=[
-                dict(key="1", summary="user story 1", url="https://jira/browse/1", points="10.0"),
-                dict(key="2", summary="user story 2", url="https://jira/browse/2", points="32.0")])
+            response, value="42", entities=[self.entity(key="1", points="10.0"), self.entity(key="2", points="32.0")])
 
 
 class JiraManualTestDurationTest(JiraTestCase):
     """Unit tests for the Jira manual test duration collector."""
 
+    METRIC_TYPE = "manual_test_duration"
+
     async def test_duration(self):
         """Test that the duration is returned."""
-        metric = dict(type="manual_test_duration", addition="sum", sources=self.sources)
         test_cases_json = dict(
-            issues=[
-                dict(key="1", id="1", fields=dict(summary="test 1", field=10)),
-                dict(key="2", id="2", fields=dict(summary="test 2", field=15)),
-                dict(key="3", id="3", fields=dict(summary="test 3", field=None))])
-        response = await self.collect(
-            metric, get_request_json_side_effect=[self.fields_json, test_cases_json])
+            issues=[self.issue(key="1", field=10), self.issue(key="2", field=15), self.issue(key="3", field=None)])
+        response = await self.get_response(test_cases_json)
         self.assert_measurement(
             response, value="25",
-            entities=[
-                dict(duration="10.0", key="1", summary="test 1", url="https://jira/browse/1"),
-                dict(duration="15.0", key="2", summary="test 2", url="https://jira/browse/2")])
+            entities=[self.entity(key="1", duration="10.0"), self.entity(key="2", duration="15.0")])